Move store admin under /admin
