+++ app/Http/Controllers/Admin/Store/AddressController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Controllers\Store\Admin;
+namespace App\Http\Controllers\Admin\Store;
 
-use App\Http\Controllers\Controller;
+use App\Http\Controllers\Admin\Controller;
 use App\Models\Store;
 use Auth;
 use Request;
@@ -11,19 +11,8 @@ class AddressController extends Controller
 {
     protected $section = 'storeAdmin';
 
-    public function __construct()
-    {
-        $this->middleware('auth');
-
-        return parent::__construct();
-    }
-
     public function update($id)
     {
-        if (!Auth::user()->isAdmin()) {
-            abort(403);
-        }
-
         $address = Store\Address::findOrFail($id);
         $address->unguard();
         $address->update(Request::input('address'));
