+++ app/Http/Controllers/Admin/Store/OrderItemController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Controllers\Store\Admin;
+namespace App\Http\Controllers\Admin\Store;
 
-use App\Http\Controllers\Controller;
+use App\Http\Controllers\Admin\Controller;
 use App\Models\Store;
 use Auth;
 use Request;
@@ -10,17 +10,6 @@ use Request;
 class OrderItemController extends Controller
 {
     protected $section = 'storeAdmin';
-
-    public function __construct()
-    {
-        $this->middleware('auth');
-
-        if (Auth::user() && !Auth::user()->isAdmin()) {
-            abort(403);
-        }
-
-        return parent::__construct();
-    }
 
     public function update($orderId, $orderItemId)
     {
